+++ packages/component/src/Attachment/ImageAttachment.js
@@ -3,18 +3,8 @@ import React from 'react';
 
 import ImageContent from './ImageContent';
 
-const ImageAttachment = ({ activity, attachment }) => {
-  const { attachmentThumbnails } = activity.channelData || {};
-
-  if (attachmentThumbnails) {
-    const attachmentThumbnail = attachmentThumbnails[activity.attachments.indexOf(attachment)];
-
-    if (attachmentThumbnail) {
-      return <ImageContent alt={attachment.name} src={attachmentThumbnail} />;
-    }
-  }
-
-  return <ImageContent alt={attachment.name} src={attachment.contentUrl} />;
+const ImageAttachment = ({ attachment }) => {
+  return <ImageContent alt={attachment.name} src={attachment.thumbnailUrl || attachment.contentUrl} />;
 };
 
 ImageAttachment.propTypes = {