+++ packages/component/src/Middleware/Attachment/createCoreMiddleware.js
@@ -8,12 +8,6 @@ import UploadAttachment from '../../Attachment/UploadAttachment';
 import TextAttachment from '../../Attachment/TextAttachment';
 import VideoAttachment from '../../Attachment/VideoAttachment';
 
-function hasThumbnail({ attachments = [], channelData: { attachmentThumbnails = [] } = {} }, attachment) {
-  const attachmentIndex = attachments.indexOf(attachment);
-
-  return !!attachmentThumbnails[attachmentIndex];
-}
-
 // TODO: [P4] Rename this file or the whole middleware, it looks either too simple or too comprehensive now
 export default function createCoreMiddleware() {
   return () => next => {
@@ -21,9 +15,9 @@ export default function createCoreMiddleware() {
       activity = {},
       activity: { from: { role } } = {},
       attachment,
-      attachment: { contentType, contentUrl } = {}
+      attachment: { contentType, contentUrl, thumbnailUrl } = {}
     }) =>
-      role === 'user' && !/^text\//u.test(contentType) && !hasThumbnail(activity, attachment) ? (
+      role === 'user' && !/^text\//u.test(contentType) && !thumbnailUrl ? (
         <UploadAttachment activity={activity} attachment={attachment} />
       ) : /^audio\//u.test(contentType) ? (
         <AudioAttachment activity={activity} attachment={attachment} />